pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plan-cascade"
version = "3.3.0"
description = "Plan Cascade - Three-layer parallel development framework for AI coding tools"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Plan Cascade Contributors"}
]
keywords = [
    "planning",
    "development",
    "ai",
    "mcp",
    "claude",
    "cursor",
    "orchestration",
    "multi-agent"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
    "keyring>=24.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
llm = [
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
]
all = [
    "plan-cascade[dev,llm]",
]

[project.scripts]
plan-cascade = "plan_cascade.cli.main:main"
plan-cascade-mcp = "mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/anthropics/plan-cascade"
Documentation = "https://github.com/anthropics/plan-cascade/blob/main/docs/MCP-SERVER-GUIDE.md"
Repository = "https://github.com/anthropics/plan-cascade"

[tool.setuptools.packages.find]
where = ["src"]
include = ["plan_cascade*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"*" = ["*.json", "*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501", "F401", "F841"]
